requirements.txt
openai>=1.30.0
deepgram-sdk>=3.5.0
pyaudio>=0.2.14
webrtcvad>=2.0.10
pyttsx3>=2.90
pyyaml>=6.0
python-dotenv>=1.0.0
setuptools>=69.0.0,<81
pytest>=8.0.0
pytest-mock>=3.14.0
